fix: improve character and ship candidates searching (#9)
There's an issue with character names including or consisting of ship
names. Previously I worked around this by retrieving the longest ship
names first and then searching for characters. However specific
combinations of character name and ship could cause the character name
and ship to be intermixed and incorrectly parsed. This fix changes the
order to make it more likely to match the correct character first, and
then hopefully get the correct ship as well.

Due to the game chat logs not providing any type information or hints
about what kind of entity was linked or posted this become very
challenging and require making some accuracy sacrifices.

In the future it might be a good idea to get multiple solutions and then
score them by probability and coherence. For example; a single character
would never fly two ships at once, that fact could act as a tie breaker
when deciding which solution to choose.